requirements.txt
requests>=2.32.3
streamlit>=1.44.0
beautifulsoup4>=4.13.3
bs4>=0.0.2
fastapi>=0.115.12
joblib>=1.4.2
lxml>=5.3.1
pandas>=2.2.3
pydantic>=2.10.6
pymongo>=4.11.3
torch>=2.6.0
transformers>=4.50.1
pillow>=11.1.0